Move GarchingArchiveModel to loaders, since it actually just loads to a PinchedModel
